Move inline classes out of `spec_helper`.
